requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
APScheduler==3.11.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
bidict==0.23.1
blinker==1.9.0
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cmdstanpy==1.2.5
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.5
cssselect2==0.8.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
discord.py==2.6.4
distro==1.9.0
Django==5.2
dnspython==2.8.0
email-validator==2.3.0
eventlet==0.40.3
executing==2.2.0
fastapi==0.129.0
filelock==3.19.1
flasgger==0.9.7.1
Flask==2.2.5
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
flask-mongoengine==1.0.0
Flask-SocketIO==5.5.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.60.0
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-pasta==0.2.0
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.71.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.13.0
holidays==0.81
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.10
image==1.5.33
imageio==2.37.0
importlib_resources==6.5.2
imutils==0.5.4
iniconfig==2.3.0
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.0
joblib==1.4.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.9.2
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
lxml==6.0.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.21
mistune==3.1.4
ml_dtypes==0.5.1
mongoengine==0.29.1
mpmath==1.3.0
msvc_runtime==14.42.34433
multidict==6.7.1
namex==0.0.9
narwhals==2.7.0
networkx==3.5
numpy==2.1.3
openai==1.109.1
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.3.2
parso==0.8.4
pdfkit==1.0.0
pillow==11.3.0
plotly==6.3.1
pluggy==1.6.0
polars==1.34.0
polars-runtime-32==1.34.0
prompt_toolkit==3.0.50
propcache==0.4.1
prophet==1.1.7
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.1.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
reportlab==4.4.3
requests==2.32.5
rich==14.0.0
rpds-py==0.27.1
rsa==4.9.1
scikit-fuzzy==0.5.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.16.2
seaborn==0.13.2
sentencepiece==0.2.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
solid==0.2.0
sounddevice==0.5.1
sqlparse==0.5.3
stack-data==0.6.3
stanio==0.5.1
starlette==0.52.1
svglib==1.5.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.204
ultralytics-thop==2.0.17
uritemplate==4.2.0
urllib3==2.5.0
utils==1.0.2
uvicorn==0.41.0
Watershed==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
yarl==1.22.0
